chore: implement mobx in project features settings
