+++ web/components/project/index.ts
@@ -1,18 +1,18 @@
+export * from "./publish-project";
+export * from "./settings";
+export * from "./card-list";
+export * from "./card";
 export * from "./create-project-modal";
 export * from "./delete-project-modal";
-export * from "./sidebar-list";
-export * from "./settings-sidebar";
-export * from "./single-integration-card";
-export * from "./sidebar-list-item";
+export * from "./delete-project-section";
+export * from "./form-loader";
+export * from "./form";
+export * from "./join-project-modal";
+export * from "./label-select";
 export * from "./leave-project-modal";
 export * from "./member-select";
 export * from "./members-select";
-export * from "./label-select";
 export * from "./priority-select";
-export * from "./card-list";
-export * from "./card";
-export * from "./join-project-modal";
-export * from "./form";
-export * from "./form-loader";
-export * from "./delete-project-section";
-export * from "./publish-project";
+export * from "./sidebar-list-item";
+export * from "./sidebar-list";
+export * from "./single-integration-card";
