+++ web/components/project/settings/index.ts
@@ -1,0 +1,1 @@
+export * from "./features-list";
